refactor: extract transaction mapping logic into a separate utility function
