+++ src/modules/safe-shield/safe-shield.service.ts
@@ -13,6 +13,7 @@ import type {
   TransactionData,
 } from '@/modules/safe-shield/entities/transaction-data.entity';
 import { ILoggingService, LoggingService } from '@/logging/logging.interface';
+import { mapDecodedTransactions } from './utils/transaction-mapping.utils';
 
 /**
  * Main orchestration service for Safe Shield transaction analysis.
@@ -71,7 +72,7 @@ export class SafeShieldService {
    * @param data - The transaction data
    * @returns Array of transaction objects with operation, to, value, data, and decoded information
    */
-  private async extractTransactions(
+  async extractTransactions(
     chainId: string,
     to: Address,
     data: Hex,
@@ -106,53 +107,9 @@ export class SafeShieldService {
 
     // Map the decoded transactions to a flat array
     return decodedTransactions.reduce<Array<DecodedTransactionData>>(
-      (acc, curr) => [...acc, ...this.mapDecodedTransactions(curr)],
+      (acc, curr) => [...acc, ...mapDecodedTransactions(curr)],
       [],
     );
-  }
-
-  /**
-   * Maps decoded transaction data recursively to an array of decoded transactions.
-   *
-   * @param args - The decoded transaction data
-   * @returns Array of decoded transactions
-   */
-  private mapDecodedTransactions({
-    data,
-    dataDecoded,
-    operation,
-    to,
-    value,
-  }: DecodedTransactionData): Array<DecodedTransactionData> {
-    if (dataDecoded?.method === 'execTransaction') {
-      const dataParam = dataDecoded.parameters?.[2];
-
-      // Recursively map the execTransaction parameters
-      return this.mapDecodedTransactions({
-        data: (dataParam?.value as Hex) ?? '0x',
-        dataDecoded: (dataParam?.valueDecoded as DataDecoded) ?? null,
-        operation,
-        to: dataDecoded.parameters?.[0].value as Address,
-        value: dataDecoded.parameters?.[1].value as string,
-      });
-    }
-
-    if (
-      this.multiSendDecoder.helpers.isMultiSend(data) &&
-      Array.isArray(dataDecoded?.parameters?.[0].valueDecoded)
-    ) {
-      // Recursively map the multiSend inner transactions
-      return dataDecoded.parameters?.[0].valueDecoded.flatMap((tx) =>
-        this.mapDecodedTransactions({
-          ...tx,
-          data: tx.data ?? '0x',
-          dataDecoded: tx.dataDecoded as DataDecoded,
-        }),
-      );
-    }
-
-    // Return the decoded transaction data if it's not a multiSend or execTransaction
-    return [{ data, dataDecoded, operation, to, value }];
   }
 
   /**
